% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{to_matrix}
\alias{to_vector}
\title{Convert data.frame to various format}
\usage{
to_matrix(.data)

to_vector(.data)
}
\arguments{
\item{.data}{A tbl.}
}
\value{
An object (matrix, vector etc.).
}
\description{
Convert data.frame to various format like matrix, vector
}
