% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpurifyr.R
\docType{package}
\name{dpurifyr-package}
\alias{dpurifyr}
\alias{dpurifyr-package}
\title{dpurifyr: a practical way for data preprocessing}
\description{
dpurifyr package gives you a practical way for `data preprocessing`
providing a consistent set of verbs that
help you solve the most common data preprocessing challenges.
}

