% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_count}
\alias{encode_count}
\title{Replace labels with their count in the data set}
\usage{
encode_count(.data, ...)
}
\arguments{
\item{.data}{A tbl.}

\item{...}{One or more unquoted expressions separated by commas. You can treat variable names like they are positions.
It supports [dplyr::select]  comparable.}
}
\value{
An object of the same class as .data with "preprocessing-chain" attribution.
}
\description{
Simply converting each values (character or factor) in a column to the count of each values
}
\details{
It can be sensitive to outliers.
It might be a good strategy to add log-transformation
}

