% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_label}
\alias{encode_label}
\title{Encode labels with value}
\usage{
encode_label(.data, ...)
}
\arguments{
\item{.data}{A tbl.}

\item{...}{One or more unquoted expressions separated by commas. You can treat variable names like they are positions.
It supports [dplyr::select]  comparable.}
}
\value{
An object of the same class as .data with "preprocessing-chain" attribution.
}
\description{
Simply converting each values (character or factor) in a column to a specific number.
}
\details{
The mapped value start from 1 (not 0).
}

