% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{print.preprocessed_data}
\alias{print.preprocessed_data}
\title{Printing Preprocessing object}
\usage{
\method{print}{preprocessed_data}(x, ...)
}
\arguments{
\item{x}{object of class data.frame.}

\item{...}{optional arguments to print or plot methods.}
}
\description{
Print a preprocessed object.
}
