% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{reexports}
\title{Automatically import `%>%` opeator when call library("dpurifyr")}
\description{
Automatically import `%>%` opeator when call library("dpurifyr")
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}
\keyword{internal}

