% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{scale.data.frame}
\alias{scale.data.frame}
\title{Get JPGIS2.1 Data}
\usage{
\method{scale}{data.frame}(.data, ..., center = TRUE, scale = TRUE)
}
\arguments{
\item{.data}{}
}
\description{
Get JPGIS2.1 Data
}

