% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_datetime}
\alias{split_datetime}
\title{Split date object into 8 components(year, month, day, the day of week, hour, minute, second and timezone)}
\usage{
split_datetime(.data, ..., sep = "_", format = NULL)
}
\arguments{
\item{.data}{A tbl.}

\item{...}{One or more unquoted expressions separated by commas. You can treat variable names like they are positions.
It supports [dplyr::select] comparable.}

\item{sep}{Separator between columns.}

\item{format}{a character string of formats.
It should include all the separators and each format must be prefixed with argument of strptime.}
}
\value{
An object of the same class as .data with "preprocessing-chain" attribution.
 <prefix>weekday columns is the day of the week as a number (1-7, Sunday is 1).
}
\description{
Split date object into 8 components(year, month, day, the day of week, hour, minute, second and timezone)
to a format that works better when you want to specify the effect of year, month, day, the day of week, hour, minute, second or timezone.
}

