% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_standard}
\alias{scale_standard}
\title{Standardize each columns by removing the mean and scaling to unit standard deviation}
\usage{
scale_standard(.data, ..., center = TRUE, scale = TRUE)
}
\arguments{
\item{.data}{}
}
\description{
Centering and scaling happen independently on each feature by computing the relevant statistics on .
Mean and standard deviation are then stored to be used on later data using the transform method.
}

